break up AppsPanel into smaller components
